+++ snapshots/checkout.go
