+++ snapshots/snapshot.go
@@ -2,6 +2,8 @@ package snapshots
 
 const PACK_ID_LEN int = 64
 
+var VerboseMode bool
+
 type Snapshot struct {
 	Message      string
 	Time         string
