+++ snapshots/util.go
@@ -1,0 +1,7 @@
+package snapshots
+
+func Check(e error) {
+	if e != nil {
+		panic(e)
+	}
+}
