--- src/options/go.mod
@@ -1,3 +1,0 @@
-module github.com/akbarnes/gover/src/options
-
-go 1.16
