--- src/options/options.go
@@ -1,3 +1,0 @@
-package options
-
-var VerboseMode bool
