--- src/util/go.mod
@@ -1,3 +1,0 @@
-module github.com/akbarnes/gover/src/util
-
-go 1.16
